+++ example/src/main/java/com/moengage/segment/example/SampleApplication.java
@@ -3,6 +3,11 @@ package com.moengage.segment.example;
 import android.app.Application;
 import com.moengage.core.LogLevel;
 import com.moengage.core.MoEngage;
+import com.moengage.core.config.GeofenceConfig;
+import com.moengage.core.config.LogConfig;
+import com.moengage.core.config.NotificationConfig;
+import com.moengage.core.internal.logger.Logger;
+import com.moengage.core.model.IntegrationPartner;
 import com.moengage.firebase.MoEFireBaseHelper;
 import com.moengage.inapp.MoEInAppHelper;
 import com.moengage.pushbase.MoEPushHelper;
@@ -16,9 +21,12 @@ import com.segment.analytics.android.integrations.moengage.MoEngageIntegration;
  * Date: 2020/09/02
  */
 public class SampleApplication extends Application {
+
+  private static final String TAG = "SampleApplication";
+
   @Override public void onCreate() {
     super.onCreate();
-
+    Logger.v(TAG + " onCreate() : ");
     Analytics analytics = new Analytics.Builder(this, BuildConfig.SEGMENT_WRITE_KEY)
         .use(MoEngageIntegration.FACTORY)
         .build();
@@ -26,22 +34,22 @@ public class SampleApplication extends Application {
 
     MoEngage moEngage =
         new MoEngage.Builder(this, "")//enter your own app id
-            .enableLogs(LogLevel.VERBOSE)//enabling Logs for debugging
-            .enableLogsForSignedBuild() //Make sure this is removed before apps are pushed to
-            // production
-            .setNotificationSmallIcon(
-                R.drawable.icon)//small icon should be flat, pictured face on, and must be white
-            // on a transparent background.
-            .setNotificationLargeIcon(R.drawable.ic_launcher)
-            .enableLocationServices()//enabled To track location and run geo-fence campaigns
-            .enableMultipleNotificationInDrawer()// shows multiple notifications in drawer at one go
-            .enableSegmentIntegration()
+            //set notification data(small icon, large icon, notification color,
+            // notification tone, show multiple notifications in drawer etc..)
+            .configureNotificationMetaData(
+                new NotificationConfig(R.drawable.icon, R.drawable.ic_launcher,
+                    R.color.notificationColor, null, true, false, true))
+            //enabled To track location and run geo-fence campaigns
+            .configureGeofence(new GeofenceConfig(true, true))
+            .enablePartnerIntegration(IntegrationPartner.SEGMENT)
+            //Configure logs
+            .configureLogs(new LogConfig(LogLevel.VERBOSE, false))
             .build();
     // initialize MoEngage
     MoEngage.initialise(moEngage);
 
     // FCM event listener
-    MoEFireBaseHelper.Companion.getInstance().setEventListener(new FcmEventListener());
+    MoEFireBaseHelper.Companion.getInstance().addEventListener(new FcmEventListener());
 
     // Setting CustomPushMessageListener for notification customisation
     MoEPushHelper.getInstance().setMessageListener(new CustomPushMessageListener());